added ObjectId validation to the controllers

+++ controllers/thoughtController.js
@@ -1,11 +1,11 @@
-const { ObjectId } = require("mongoose").Types;
-const { Reaction, Thought, User } = require("../models")
+const { Thought, User } = require("../models");
+const mongoose = require("mongoose");
 
 module.exports = {
   // get all thoughts
   async getThoughts(req, res) {
     try{
-      const payload = await Thought.find(); // IDK IF THIS IS CORRECT ???????????????
+      const payload = await Thought.find();
       res.json({status: "success", payload});
       } catch(err) {
         console.log(err.message)
@@ -15,6 +15,12 @@ module.exports = {
   // Get a Thought
   async getSingleThought(req, res) {
     try{
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await Thought.findOne({_id: req.params.id});
       res.json({status: "success", payload})
     } catch(err) {
@@ -25,7 +31,15 @@ module.exports = {
   // Create a Thought
   async createThought(req, res) {
     try {
+      
       const payload = await Thought.create(req.body);
+
+      // associate thought with the user
+      await User.findOneAndUpdate(
+        { username: req.body.username}, //the Where
+        { $push: { thoughts: payload._id}} // the What
+      )
+
       res.json({status: "success", payload});      
     } catch (err) {
       console.log(err.message);
@@ -35,6 +49,12 @@ module.exports = {
   // Update a Thought
   async updateThought(req, res){
     try{
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await Thought.findOneAndUpdate(
         {_id: req.params.id}, // THIS IS THE WHERE
         { $set: req.body}, // THIS IS THE WHAT
@@ -55,9 +75,15 @@ module.exports = {
   // Delete a Thought
   async deleteThought(req, res) {
     try {
-      const payload = await Thought.findByIdAndDelete(req.params.id); // WORK???????????????????
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
+      const payload = await Thought.findByIdAndDelete(req.params.id);
       if (!payload){
-        res.status(404).json({message: "No thought with that ID"});// PROBABLY WRONG ??????????????????????????????????????????
+        res.status(404).json({message: "No thought with that ID"});
       }
       res.json({status: "success", payload})
     } catch(err) {
@@ -67,6 +93,12 @@ module.exports = {
 
   async createReaction(req, res) {
     try{
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await Thought.findOneAndUpdate(
         {_id: req.params.thoughtId}, // THIS IS THE WHERE
         { $push: {
@@ -75,7 +107,10 @@ module.exports = {
         { runValidators: true, new: true}, // new: true - returns the updated object
         // runValidators:true - enforces validators from the schema on the updated object
       );
-      console.log(req.params.thoughtId)
+
+      if(!payload){
+        res.status(404).json({message: "Thought not found!"})
+      }
       res.json({status: "success", payload})
     } catch (err){
       res.status(500).json({status: error, payload: err.mesage})
@@ -84,6 +119,12 @@ module.exports = {
 
   async deleteReaction(req, res){
     try{
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await Thought.findOneAndUpdate(
         {_id: req.params.thoughtId},
         {$pull: {
@@ -92,6 +133,11 @@ module.exports = {
           }
         }}
       )
+
+      if(!payload){
+        res.status(404).json({message: "Thought id doesn't exist or reaction not created!"})
+      }
+
       res.json({status: "success", payload})
     } catch(err){
       res.status(500).json({status: error, payload: err.mesage})
