+++ controllers/userController.js
@@ -1,10 +1,13 @@
 const { User } = require("../models")
+const mongoose = require("mongoose")
 
 module.exports = {
   // get all Users
   async getUsers(req, res) {
     try{
-      const payload = await User.find().populate([{path: "thoughts"}, {path: "friends"}]); // IDK IF THIS IS CORRECT ???????????????  NEED TO INCLUDE USER
+      const payload = await User.find()
+        .populate({path: "thoughts", populate:{path: "reactions"}})
+        .populate("friends"); 
       res.json({status: "success", payload});
       } catch(err) {
         console.log(err.message)
@@ -14,7 +17,15 @@ module.exports = {
   // Get a User
   async getSingleUser(req, res) {
     try{
-      const payload = await User.findOne({_id: req.params.id}).populate([{path: "thoughts"}, {path: "friends"}]);
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
+      const payload = await User.findOne({_id: req.params.id})
+        .populate({path: "thoughts", populate: {path: "reactions"} })
+        .populate("friends");
       res.json({status: "success", payload})
     } catch(err) {
       console.log(err.message)
@@ -34,6 +45,12 @@ module.exports = {
   // Update a User
   async updateUser(req, res){
     try{
+      // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+      if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await User.findOneAndUpdate(
         {_id: req.params.id}, // THIS IS THE WHERE
         { $set: req.body}, // THIS IS THE WHAT and "$set" - only changes things that are changed
@@ -54,9 +71,16 @@ module.exports = {
   // Delete a User
   async deleteUser(req, res) {
     try {
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await User.findByIdAndDelete(req.params.id); // WORK???????????????????
       if (!payload){
-        res.status(404).json({message: "No user with that ID"});// PROBABLY WRONG ??????????????????????????????????????????
+        res.status(404).json({message: "No user with that ID"});
+        return
       }
       res.json({status: "success", payload})
     } catch(err) {
@@ -66,18 +90,22 @@ module.exports = {
   // CREATE FRIEND
   async createFriend(req, res) {
     try{
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await User.findOneAndUpdate(
         {_id: req.params.id},
-        {$push: {
-          friends: req.params.friendId
-        }},
+        {$push: {friends: req.params.friendId} },
         {runValidators: true, new: true}
       )
 
       if (!payload) {
         return res.status(404).json({ message: "User not found." });
       }
-      
+
       res.json({status: "success", payload})
     } catch(err) {
       res.status(500).json({status: "error", payload: err.message})
@@ -85,6 +113,12 @@ module.exports = {
   },
   async deleteFriend(req, res) {
     try {
+       // check if the provided ID is a valid ObjectId (could simply be missing a digit)
+       if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+        res.status(400).json({message: "Invalid ID format"})
+        return
+      }
+
       const payload = await User.findOneAndUpdate(
         {_id: req.params.id},
         {$pull: {friends: req.params.friendId} },
@@ -100,7 +134,6 @@ module.exports = {
       res.status(500).json({status: "error", payload: err.message})
     }
   }
-  //addFriend deleteFriend and i only need to _id and "friends" _id in the req. params
-  // http://localhost:3001/api/user/656921fbbc49b7e17d07aa9e/friend/65692780d88c52d9be5f50f0
+
 }
